Migrating Drydocks to use Fleet. Bugs remain.
NameFactory has become a singleton-like class, with a single object
for the entire program to use.
There is a bug where purchasing a new ship will not update the Right
screen.
There is a bug where you can purchase more ships than there are slots (4)
There is a bug where the 'page dots' do not appear.